Reduce typing for local binstub commands
Previously we recommended using the local bundler binstubs to avoid the
use of `bundle exec` when running commands in the context of the local
bundle. This adds the local bin directory to the PATH for most services,
so that the local bundler binstubs are used automatically - without
needing to type 'bin/' - and in preference to any system binaries.