Move E2E whitehall setup task into master Makefile
Previously the Whitehall Makefile had a task to populate topic data,
which relied on the E2E stack. Running the task in the specific Makefile
for Whitehall introduced a potential ordering issue, since all the
dependencies for Whitehall must be setup in order to run its E2E stack.

This moves the task into the master Makefile, which seems reasonable
since it actually does setup across multiple services.