Remove build-essential everywhere
Previously we installed the build-essential package for all images,
which is an artefact from the approach in the E2E tests. This removes
the package, since the utilities it provides (make, gcc, etc.) are
included as part of the ruby:* images.

This also removes an unnecessary set of lib* packages for publishing-api
which are again an artefact from copying the E2E test Dockerfiles.