Remove 'fleshclam' command from Dockerfile
Previously we ran 'freshclam' to update virus definitions for Asset
Manager, but this takes a while a doesn't seem to have any impact on
the day-to-day tasks. This removes the command from the Dockerfile -
we may want to support updating virus definitions more regularly and
easily if this is a day-to-day concern.