Trial a few simple build tests
This adds a few no-cache (from-scratch) build commands for services with
different Dockerfiles, to test the builds continue to work over time as
external sources change; it should also help to double-check any global
changes made across the Dockerfiles.

Running from-scratch builds can take a while. One way to mitigate this
is be to use the '--parallel' flag, which runs the builds in parallel at
the expense of making the output jumbled - this seems like an OK
compromise as we expect the commands to be successful most of the time.

Currently takes around 2 mins using the 'time' command, and around 3.2
mins without the '--parallel' flag.