Create draft content store db
This also seeds the db with a user which is required to use the draft
content store.

While setting up calendars we ran into an issue due to there being no
user existing in the draft-content-store. The calendar publishing rake
task
(https://github.com/alphagov/calendars/blob/master/lib/tasks/publishing_api.rake)
was failing and the draft content store wasn't being populated with the
calendars. This commit adds a user for the draft-content-stores db when
running `make` for content-store.

Trello:
https://trello.com/c/sm2rw3Nr/27-set-up-calendars