Use binstubs in all Makefiles
Previously we invoked commands like 'rails' and 'rake' in the Makefiles
outside of the bundler context, which can lead to confusing errors. This
introduces the use of binstubs - i.e. 'bin/rake' and 'bin/rails' - in
all Makefiles to ensure we run commands in the context of the local bundle.