Add doctor functionality for dnsmasq
This is currently only tested on a Mac.
Add some basic tests to help diagnose issues with dnsmasq:

* Test whether dnsmasq is installed
* Test whether dnsmasq is running

Print out some helpful messages to reflect the state to the user.
Include instructions on remedial action if needed.

Redirect the system calls to /dev/null so we don't print to STDOUT when
running the command.