Rename 'run-this' to 'run'
Now that we're no longer allowing arbritary docker-compose commands we
no longer have a clash with the docker-compose run command so we can
rename 'run-this' to 'run' meaning devs have to type less.

I've allowed the old 'run-this' command to exist for backwards
compatibility reasons but it will be removed at some point in the
future.