Create generic Doctor Checkup class
We have a few classes now that do very similar things when running
`govuk-doctor`. This is an attempt to make a more generic solution and
DRY up the code a little. It will allow for a series of optional
checkups based on an array of symbols passed in and will return messages
based on custom messages passed in as a hash.