Rename nginx-proxy container to nginx-proxy-app
Currently govuk-docker assumes all containers have the format
`#{service}-#{stack}` and this is not the case for nginx-proxy.

Rather than change govuk-docker to support this special case, I thought
it would be better to rename `nginx-proxy` for consistency with our
other apps. This allows you to run it using `govuk-docker startup`.