Add elasticsearch dependencies
Search API is currently running with both Elasticsearch 5 and
Elasticsearch 6 while a migration is undertaken to move to 6.

One thing to note is the environment variable `ES_JAVA_OPTS=-Xms512m -Xmx512m`
- Docker has a default memory size of 2GB and Elasticsearch also has a
default memory size of 2GB. When attempting to bring up both, the container
runs out of memory and one gets killed so this prevents that.