Add Search API web app
Before Search API can run it needs indices created, thus these are added
as a task in the makefile. As they are needed for both Elasticsearch 5
and Elasticsearch 6 these are both added as dependencies of the lite
stack.

A consequence of this is that tests won't pass against lite unless an
environment variable is specified to kill off the Elasticsearch B
instance.