Update readme for publisher
I was tempted to flag this with a warning as we haven't set up
Imminence, however I can't find any reference in the code that actually
uses Imminence outside of tests.

Publisher seems to have a problems where there are dependencies that may no
longer be needed defined in the readme. Another is asset-manager which
seems to only be used by some models that don't seem to be part of
this application.