Add explicit build step to Makefiles
Previously we relied on docker-compose to auto-build each service image
at the point when it is first run. If a Dockerfile changes, we need to
support users in rebuilding their images, which docker-compose will not
do by default. This adds an explicit 'compose build app-name-lite' step
to each service Makefile to help ensure this happens.

Rather than repeating service names over and over, this also replaces
the service name with '$@', which strongly coupled the name of the
service to the Makefile target. At the same time, it didn't make much
sense to add another command variable to do the build, so this replaces
COMPOSE_RUN with explicit use of govuk-docker compose run, alongside
govuk-docker compose build.