Add a paths helper class
More and more commands are starting to need to know where govuk-docker
is so I thought it would be useful to pull that out into a class. By
doing this we can be consistent with the use of environment variables
and make sure they match the ones used by the Makefile.