Fix makefile dependencies and bundler versioning
Previously we installed bundler without a version restriction, which
caused problems since Bundler 2.0 is not compatible with most of our
repos. We may also need to cope with installing specific versions of
bundler for certain services. This splits the installation step for
bundler and ensures the major version is consistent with each service.

Previously we used '$@' in the target list for each service, which is
not valid (it can only be used within the recipe). This reverts the use
of '$@' back to the service name in each service Makefile.