Fix `govuk-docker doctor` false positive
I initially thought that
`system "git diff master origin/master | cat 1>/dev/null"`
returned false if there were no differences and true if there were,
so I wrote the `up_to_date?` method accordingly. This is not true.
The tests passed because they were written under this assumption.

What we can use instead is the `--exit-code` flag that makes the program
exit with 0 if there are no differences and 1 if there are differences.

The `--quite` flag must be used instead of `| cat 1>/dev/null` to ensure
that the return code is the correct one and not the `cat`'s return code.

I know that `--quiet` implies `--exit-code` but I think that having both
is clearer and more explicit.