Rename 'clone' target to make it consistent
Previously we used '../%' as the target for cloning service '%'. The
rule also had a guard clause to check the service exists, but this seems
unnecessary since there is nothing to suggest people will invoke the
rule manually. This renames the target to 'clone-%' to make it consistent
with the new 'bumdle-%' target and more self-explanatory, to the point
where the comment for the rule seems unnecessary.