Update pre-requisites
'govuk-docker setup' handles installing Docker, docker-compose
and Dnsmasq, so we don't need to get the developer to manually
install those. But we should make it clear that they need a
Mac machine to follow these instructions properly, and that
installation is dependent on Ruby and Git.

See https://github.com/alphagov/govuk-docker/pull/141/files\#r301681626