Focus the 'Usage' section on the two main commands
Previously the 'Usage' section had a lot of different command
variations, which could be confusing for a first-time user. This
modifies the section to concentrate on the two main commands: run
(tests) and startup in just one way to keep things simple.

This also moves the supplementary information about the different stacks
into the 'Stacks' section, since the use of alternate stacks is quite
advanced and it makes more sense to show the commands after explaining
what a stack is and the different stacks available.