Add healthchecks for backing services
By adding a healthcheck we ensure that containers which rely on these
services don't come up until the service reports a healthy state.
e.g. search-api won't be available to do things before elasticsearch
is running.  This is useful because at the moment there is a race
condition in commands like this:

    GOVUK_DOCKER_SERVICE=search-api \
    govuk-docker run curl -XDELETE http://elasticsearch5:9200/_all

As the containers are started at the same time, without waiting, there
is a good chance the `curl` will happen before Elasticsearch is able
to respond to requests.  A healthcheck fixes this.

These checks were copied from the e2e tests.