Add a "replication" volume to ES containers
This is so we can download an ES snapshot from integration, copy it
into the container, and then restore the snapshot from the local
filesystem.  Restoring directly from s3 is undesirable because these
snapshots can be several gigabytes, and if the connection drops during
the restore it has to be started again.

The ES configuration files are needed because the path.repo option
can't be set by environment variables.