Create directory before writing file
If a directory does not exist before writing a file, setup will error
out.

Add function to create the directory if it does not yet exist.

Since one of the directories we need to create is /etc/resolver we need
to allow for creating the directory as `sudo`. To allow for this, add
`as_sudo` as a parameter to existing `write_file`. This defaults to
false.