Add MySQL command to increase max allowed packet size
Importing replicated data into Whitehall's database has proved problematic as
the default max_allowed_packet value for MySQL isn't large enough. Previously
we tried adding max_allowed_packet as an environment variable but this didn't
work. Instead we are passing it as a command argument, noting that the
entrypoint file for MySQL looks for this argument and prepends it with
"mysqld".