Add required frontend dependency for Whitehall
Previously we could not run the E2E setup task for Whitehall since the
homepage content item that it relies on did not exist. This adds a
dependency on the frontend app, which includes a setup task to publish
special routes/content, which includes the homepage.

This also adds the 'frontend' app as a dependency for Content Tagger,
since it too relies on the existence of the homepage content item.