Document how to remove Docker containers
Myself and a few members of the step by step team have had to resort to this a few times
when working with collections-publisher, which would sometimes get in a painful state trying
to interact with publishing-api or link-checker-api.  It did not appear to be fixable through the
normal methods of updating and rebuilding individual services, running in an `e2e` stack or
restarting Docker itself.

The most reliable way to fix some of the more obscure problems has been to clear all govuk
containers and rebuild them. The commands are not currently documented in the README.