Fix draft static referring to live static assets
Previously we depended on both the draft (draft-static.dev.gov.uk) and
app (static.dev.gov.uk) stacks of static when starting government-frontend,
since some of the network requests looked like they were hard-coded to use
the app stack. Further investigation shows this assumption was incorrect,
and that the draft stack of the static app can be corrected to correctly
refer to itself for asset URLs in the templates it renders.