Deprecate the 'prune' command
Previously we had a 'prune' command to cleanup containers, dangling
images and volumes. This is usually used in the context of stopping
and cleaning-up containers specifically, which is not fully covered
by this command.  Rather than extend the 'prune' command further,
this deprecates it in favour of a more standard 'compose' command,
which addresses the specific need of stopping/cleaning containers.