Remove auto-env feature
Previously each 'run' command allowed the user to assign environment
variables by automatically wrapping user command in 'env'. This replaces
the automated use of 'env' with a how-to article, so that the technique
is better documented, while having a lower code footprint for the
limited number of occasions when it is actually used.