Remove hostname printer
Previously we added a helper thread to indicate where a started app
would be available. Initially the output from this thread confused me,
as it looked like it was coming from the app itself, as opposed to
govuk-docker. Despite rescuing an increasing number of exceptions,
it also still seems to error for me, or print output prematurely.
This removes the helper thread, on the basis that most domains should be
intuitive if the user has previously visited one for another app.