Align better with docker-compose for startup cmd
Previously we delegated to 'docker-compose run' when attempting to
'startup' an app, which is at odds with the conventional use of the
'docker-compose up' command for this purpose. Our over-use of 'run'
lead to some confusion:

   - 'run' recommends out the 'lite' stack when invoked via 'startup'
with a typoed variant
   - 'run' has a '--service-ports' flag, which only applies when
invoked via 'startup'

This switches the 'startup' command to do a 'docker-compose up',
which removes the need for the (now implicit) '--service-ports' flag,
and the point of confusion around recommending a 'lite' stack for
startup. Instead, the documentation now has more explanation.