Merge and extract installation to a new doc
Even after moving the compatibility section to a new doc, I'm still
finding the volume of the README to be daunting and hard to follow.
This tries to address that further by combining a bunch of related
and volumous sections into a new 'installation' doc - having relevant
how-to's closer to the setup instructions should also make them more
discoverable when things do go wrong.