Auto-cleanup pidfile when rails starts
Previously we tried to use the container's /tmp directory to auto
cleanup the Puma pidfile when each app quits. Ideally Puma would be
able to cleanup the file itself, but sometimes the Docker process
exits too quickly for this to happen, which can cause future attempts
to start the server to error with 'A server is already running.'.

Switching to the docker-compose 'up' command has caused the '/tmp'
technique to stop working, since the web app container is no longer
destroyed when the process quits. This replaces the '/tmp' technique
with the rails server '--restart' flag, which does the necessary
actions to cleanup the pidfile before the server tries to start.

The '--restart' flag is primarily intended for use with the 'rake
restart' task, which touches 'tmp/restart.txt', which causes Puma
to invoke 'rails server' with this '--restart' flag. When a container
fails to exit cleanly, it seems reasonable to treat a subsequent
attempt to start it again as a restart, since it never truly stopped.