Add mapit
This adds Mapit as a service with a dependency on PostgreSQL and
nginx-proxy-app.

Mapit has a PostgreSQL database with the PostGIS extension installed to
enable location queries to be run in SQL, that's why we used a
docker image that includes this extension.

This should enable Mapit to work with govuk-docker but there will be
further work to enable importing data. There's a problem with loading
static files which is also an issue in the dev VM. However, this
shouldn't be a problem since we only really care about the database
being up to date in this app.

Co-authored-by: Ken Tsang <ken.tsang@digital.cabinet-office.gov.uk>