Fix 'require' statements using LOAD_PATH
Previously we used 'require_relative' statements in code and tests,
which all broke in the previous commit due to the directory change.
This replaces the use of 'require_relative' everywhere by modifying
the LOAD_PATH to enable a single, top-level 'require' of 'govuk_docker'.

While it may not always be advisable to 'require' all code by default,
the impact here should be minimal, since the number of files is small
and we don't do any heavy work as part of loading each file. Requiring
everything by default seems like a good compromise to reduce repetition.