Use 'expose' option instead of 'ports'
Previously we used the 'ports' option to declare the internal ports
for app services. In order for the declared ports to be exposed to the
internal nginx-proxy service, it was necessary to use the 'up' command
or the '--service-ports' flag with the 'run' command. Using the 'expose'
option means 'up' and 'run' can be used interchangeably, thus removing
a potential source of confusion.