Move nginx-proxy service to top-level
nginx-proxy has been treated like any other service since the start of
this repo, but it's actually quite different because it has no stack and
no Makefile. It's also a top-level service that's depended on by all of
the app stacks. This moves it to the top-level docker-compose.yml file
to make it more visible and simplify writing tests for (other) services.