DRY-up helper code in integration tests
This extracts a bunch of repeated or complicated code into a new set of
helpers. Since we use the helpers to define the specs, as well as part
of running them, we need to call the helpers explicitly when using them,
since there's no clean way to 'include' them at all levels of the code.