Fix postgres database replication filename pattern
At some point the filename pattern changed: backups are now in the
"postgresql-backend" directory, and it's not consistent at all whether
things use hyphens or underscores.

I decided to use `aws s3 ls ... | grep` to find the app, rather than
hard code a list of app name -> backup file name mappings.  The list
would be long, and the `aws s3 ls` is needed anyway, as the backup
timestamps also now include the time the backup was taken, not just
the date.