Rewrite setup script to be a single file
Previously we had a Ruby script to setup GOV.UK Docker, which was hard
to understand because the setup spanned multiple files and methods.

This rewrites the setup into a single script. A couple of features of
the original script haven't been carried over:

   - Since all the changes are necessary and isolated, we don't ask
before doing them
   - We only expect people to run setup once in the repo, so no more
repo/pull stuff

This also adds an extra command to ensure the system resolver picks up
the new config in /etc/resolver, which has been a problem for some.