Replace overloaded 'service' term with 'project'
Previously we used the word 'service' to describe the combination of
apps and libraries that make up GOV.UK. Unfortunately this is an
overloaded term: Docker Compose also uses the term 'service' to describe
something you can start/run. In this terminology, each GOV.UK service
had many Docker Compose services for each of its stacks. Confusing!

This renames all uses of the term '(GOV.UK) service' with 'project' as
a way to distinguish what we are referring to. I couldn't think of a
better word to describe 'apps + libraries' in a way that makes sense
when 'running' them, 'building' them, 'starting' them, etc.