Whitehall specific Dockerfile
This creates a Dockerfile that is used for Whitehall builds. It adds the
ghostscript and mysql-client dependencies. Whitehall uses ghostscript as
a means to create image files from PDFs and it calls out to the MySQL
client during tests to test a sanitisation script.

These seem to be the missing steps needed to get a fully passing
Whitehall test suite passing with GOV.UK docker.