Automatically acquire new Ruby versions
Previously we had documentation for when an app started using
a new version of Ruby that was not in the list that's built into
the associated image. This replaces the documentation with an
automatic command to rebuild the image with the latest list of Ruby
versions. (Pointing people to docs over and over gets old fast!)