Add ASSET_HOST env var to frontend apps
This adds environment variables for all the frontend apps that can be
used to specify the hostname used for assets. This is to support the
change of hostname for assets of GOV.UK apps from the
`assets.publishing.service.gov.uk` hostname to the `www.gov.uk` one.
For more information on this change see: https://github.com/alphagov/govuk-puppet/pull/10360.

This change pre-emptively applies this environment variable to all
frontend apps with the expectation that the apps will gradually support
this env var as the apps are migrated. Once each app has been
migrated the GOVUK_ASSET_ROOT environment variables for them can be
removed.