Add missing imagemagick dependency to the Whitehall Dockerfile
Whitehall uses the mini_magick Ruby library, and I believe this
Dockerfile is written with the intention of using bundler/rubygems to
install mini_magick.

Unfortunately, bundler/rubygems aren't capable enough to sort out
installing imagemagick, which is a problem, as it's needed to make
mini_magick work. Therefore, install the Debian package for
imagemagick through the Dockerfile.