Add issue template
Previously we have received many non-specific comments about issues
with GOV.UK Docker. It's unclear if some of these are due to a lack
of understanding, or represent actual bugs to fix in this repo. For
any problem, we want to encourage people to raise an issue, so that
we can capture the discussion more permanently than on Slack. This
adds an issue template to ensure we capture sufficient detail when
someone raises an issue - enough to suggest a fix.