Add repos that relate to live GOV.UK
This adds repos to the compatibility list with the view that this
repo should support development on projects that are used to run
the live site. While it would be possible to use a local Ruby setup
to support some of these projects, we want to avoid a situation
where developers have to know when to use GOV.UK Docker / not.

Note this removes the "govuk-cdn-config" project, since the code in
the repo is not used to run the live site.