Remove all-apps from Makefile
The usefulness of this command has declined over time, and since we moved the generic ruby libraries to their own directory this command will fail as it tries to build based on folder name.

As this is not a command most users might be expected to run, we're removing it here.

If you have a need to build your own apps, you'd be better off doing so by running your own command to list relevant directories.

The following takes the existing command and does an inverted grep to create a blocklisti that could prove useful
'ls projects/*/Makefile | xargs -L 1 dirname | xargs -L 1 basename | grep -v generic-ruby-library'