Remove unused issue template
This was added in the spirit of recording issues people have here,
but that hasn't happened in practice. If we start seeing a lot of
issues created for this repo, then that would be a more sensible
point to add a template, if we're wanting some structure.

This removes the template until we actually need it.