Rewrite usage section of the README
This combines the explanation of stacks with the usage suggestions,
so that someone new to this repo learns about both in tandem. We've
seen some anecdotal evidence that people start using GOV.UK Docker
without appreciating the concept of "stacks".

Note that this also DRYs-up the repetition of "make" in the usage
guidelines. Anecdotal evidence also suggests people are running
"make" religiously for things like installing new versions of gems.
This removes ambiguity about when to run "make".