Create app-account stack for finder-frontend
Finder frontend is currently being used as part of the trial / prototype
of GOV.UK accounts. These integrations are likely to be tested with
other apps in the near future.

So here we create a re-usable new stack app-account, this will:
- spin up dependent apps required to make account features work
- provide all environment variables (eg: keys, feature flags) to ensure
things run smoothly.

For finder frontend, that adds the additional dependencies of
email-alert-api (used to get subscription titles to send to the account
in the JWT) and the account manager prototype itself.

There are also a range of other keys, IDs and environment variables
required to make the account features work.

As we now have a live stack, these variables can be removed from other
stacks where things should continue to function, but with no expecation
account features will function