Move troubleshooting info into a single doc
Previously troubleshooting guidance was split across the README and
the installation doc, but it could be relevant in several contexts.
This centralises the guidance in a single file, which also allows
more room for extension.

Note that the guidance for "bin/doctor" is deliberately duplicated,
since it may be useful for both setup and normal usage.

Note that the guidance for upgrading Ruby is rewritten as part of the
CONTRIBUTING, since it's not relevant to normal installation / usage,
and is unfortunately required in order to run the "bin/doctor" script.
(We should aim to fix this in future.)