Restructure install / how-to guidance
Previously this guidance was also spread between the README and the
installation doc. This restructures the guidance so we have:

- Minimal installation instructions directly in the README.
- Everything else is a how-to.

Due to the growing number of how-tos, these are now in a separate
file with several, non-exhaustive pointers in the README for some
of the more common post-setup tasks. This also links to the new file
in a new, general "resources" section of the README.