Add a test that Rails apps have expected volumes
We expect all Rails apps to have a govuk volume, a root-home and one for
their temporary files. This adds a test to ensure that.

We do expect most non Rails apps to also have the root-home and govuk
volume but there are exceptions to this, most notably go apps and
mapit-pg so specifying this to Rails apps covers the places we can
confidently expect these volumes.