Test that node projects have a node modules volume
This adds a test that each service within a project that uses node has a
node_modules volume. The approach I've taken to determine a node project
is to identify npm or yarn in a Makefile, this feels pretty flakey but
for now is sufficient to identify all the node projects.

It does seem possible that a project may have some services that don't
want access to the node_modules volume but it seems unlikely while it is
part of the apps source code.