Fix accounts app not building from scratch
Previously the build failed because the "lite" service depended on "app",
which depended on the attribute "app", which caused it to try and start,
which caused a "tmp" directory to be created for it, which prevents it
cloning ater on, which prevents it from starting, causing errors back
down the chain. This removes the "app" dependency, noting that I can run
the setup without it, and the tests all pass.

In general, the "lite" stack should have the minimum dependencies in
order to run basic commands for a project [1].

[1]: https://github.com/alphagov/govuk-docker#the-lite-stack