Modify account dev URL to serve from SecureContext
The account manager prototype contains UI to allow a user to log in and
manage accounts.

As part of that login journey we have been prototyping webauthn
integration. This has revealed there are a range of secure-only browser
APIs that are only available over a HTTPs connection or in another
secure context.

We looked into adding HTTPs certificates however, an easier win seemed
to be modifying the domain to serve from a `.localhost` domain which is
regarded as a secure context[1].

This may have some advantages over self signed certificates in terms of
simplicity of setup[2], but that requires a bit more of a look.

So as a result we move the URL for only the account manager from a
`dev.gov.uk` account to a `.localhost` domain.

[1]: https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts
[2]: https://medium.com/@bramus/on-secure-contexts-in-firefox-https-for-local-development-and-a-potential-nice-gesture-by-cea4ab4903a