Add new account-api app
We now have three account-related apps in govuk-docker:

1. govuk-account-manager-prototype
2. govuk-attribute-service-prototype
3. account-api

(1) and (2) are running in production on the PaaS, and are prototypes
of the cross-government authentication and attribute storage systems,
with the GOV.UK Transition Checker as the only user currently.

Responsibility for the cross-gov side of accounts is moving to the
Digital Identity (DI) programme, so those apps will be retired as soon
as DI have viable replacements.

(3) is a GOV.UK-internal app specified in [RFC 134][] which lets
different GOV.UK microservices make use of the accounts and
attributes.

When (3) is set up, finder-frontend will talk to it, and not to the
other apps.

[RFC 134]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-134-govuk-wide-session-cookie-and-login.md