Add account-api-app-live
account-api now depends on the content-store.  To remove the need to
replicate data, this project uses the live content-store.

However, account-api still needs its own database and a local
govuk-account-manager-prototype.  Perhaps this is something to bring
up with Digital Identity?  The need for services to be able to auth in
test environments.