Switch to db:prepare from db:setup
Fixes: https://github.com/alphagov/govuk-docker/issues/491

In 1ec59d4 we simplified the command to initialise a DB for an app,
since Rails 6 requires we run it for the dev and test environments
separately. The only command swallowed any errors, which masked an
issue that 'db:setup' isn't idempotent for MySQL DBs.

This switches to 'db:prepare', which has roughly the same behaviour
as the old command (including running any seeds), but with the added
bonus of not swallowing any genuine errors. Tested with Collections
Publisher (MySQL) and Content Publisher (Postgres).