Add govuk-content-schemas-app Docker configuration
This is to enable `govuk-docker-run` to work on govuk-content-schemas,
so that we can remove the startup.sh script.
This configuration was mostly copied from the code that was DRY'd
up in 75f5986063e904bcbd862e732bd088934dd1ae30, and also a bit of
inspiration from govuk-developer-docs for getting the port binding
working correctly.