Add special case for govuk-content-schemas
Without this special case, tests fail with:

```
Failures:

  1) Make dependencies mirrors docker-compose.yml for publishing-api
     Failure/Error:
       expect(compose_dependencies(project_name))
         .to match_array(makefile_dependencies(project_name))

       expected collection contained:  ["govuk-content-schemas"]
       actual collection contained:    []
       the missing elements were:      ["govuk-content-schemas"]
     # ./spec/integration/make/dependencies_spec.rb:6:in `block (3 levels) in <top (required)>'

Finished in 0.32347 seconds (files took 0.36969 seconds to load)
1081 examples, 1 failure
```

What's happening is, now that govuk-content-schemas has an `-app` stack,
it's getting picked up by makefile_dependencies. The expectation is that
if an app needs another "app" in its Makefile, then it also needs the app
to be running when it starts, but that's not the case here. The schemas
are (currently) the one exception to this rule, because it's just the
files that matter.

See https://github.com/alphagov/govuk-docker/pull/503#discussion_r671127688