Move 'dependencies that don't run' logic into methods
This makes the top-level method easier & quicker to read.
Unfortunately the array of apps is not accessible inside the
methods unless I turn it into a constant, and then Rubocop
complains, so I've added a lint override as I think this is still
a neater separation of concerns than to bury the array inside
its own method at the bottom of the file.