Test that every Makefile dependency corresponds to a project
In a previous iteration of #503, I'd accidentally added
`govuk-content-schemas-app` to the Content Store's Makefile, and
all tests continued to pass (though `make content-store` would
fail). We can catch these kinds of errors ahead of time by adding
a test, so hopefully this saves a future developer some frustration.

Note that we have some "allowed prefixes" (`bundle-` and `clone-`),
so not _every_ Makefile dependency needs to match a project name.
This is covered in the test.

Resolves https://github.com/alphagov/govuk-docker/pull/503#pullrequestreview-708547331