Fix linting issue to avoid disabling Cop
Previously we used a constant but we can achieve the same thing with
a "let" statement, which doesn't violate any linting rules.

I think the reason the original variable stopped working is that the
reference in the test was in the same scope (of the "describe" block),
whereas the methods are run in some other scope, which appears to be
the same as that populated by the "let".

(Using a constant always worked because constants are global, hence
the linting issue.)