Document how to work through database-related `make` issues
These kinds of issues crop up fairly regularly, so it's good to
have a dedicated Troubleshooting section for them. I've
keyword-stuffed the kinds of errors developers might run into where
they'd want to follow this guidance; it isn't an exhaustive list,
but could be useful and can always be added to later.

This documentation resolves #511.