Update readme to make clear that the commands are using an example app name
Currently, the readme uses collections-publisher as an example app for explaining how to use the commands to build and run. We've had a number of folks join who are trying to use docker before they fully grasp our app setup, leading to them thinking they have to run these commands against the example app, primarily collections-publisher, specifically and not understanding why they can't run the app they want to. This change spells it out in a bit more detail that the commands being used here are for every app and depend on what app you want to build/run.